added delete method
