+++ OdeToFood.Data/IRestaurantData.cs
@@ -1,6 +1,5 @@
 ﻿using OdeToFood.Core;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace OdeToFood.Data
 {
@@ -10,60 +9,7 @@ namespace OdeToFood.Data
         Restaurant GetById(int id);
         Restaurant Update(Restaurant updatedRestaurant);
         Restaurant Add(Restaurant newRestaurant);
+        Restaurant Delete(int id);
         int Commit();
-    }
-
-    public class InMemoryRestaurantData : IRestaurantData
-    {
-        readonly List<Restaurant> restaurants;
-
-        public InMemoryRestaurantData()
-        {
-            restaurants = new List<Restaurant>()
-            {
-                  new Restaurant{ Id = 1, Name = "Hot Pizzzas!!", Location= "Maryland", Cuisine = CuisineType.Italian },
-                  new Restaurant{ Id = 2, Name = "Cinnamon Club!", Location= "London", Cuisine = CuisineType.Mexican },
-                  new Restaurant{ Id = 3, Name = "Chat Pat!!", Location= "New York", Cuisine = CuisineType.Indian },
-            };
-
-        }
-
-        public Restaurant GetById(int id)
-        {
-            return restaurants.SingleOrDefault(restaurants => restaurants.Id == id);
-        }
-
-        public IEnumerable<Restaurant> GetRestaurantsByName(string name = null)
-        {
-            return from r in restaurants
-                   where string.IsNullOrEmpty(name) || r.Name.StartsWith(name)
-                   orderby r.Name
-                   select r;
-        }
-
-        public Restaurant Update(Restaurant updatedRestaurant)
-        {
-            var restaurant = restaurants.SingleOrDefault(r => r.Id == updatedRestaurant.Id);
-            if (restaurant != null)
-            {
-                restaurant.Name = updatedRestaurant.Name;
-                restaurant.Location = updatedRestaurant.Location;
-                restaurant.Cuisine = updatedRestaurant.Cuisine;
-            }
-
-            return restaurant;
-        }
-
-        public int Commit()
-        {
-            return 0;
-        }
-
-        public Restaurant Add(Restaurant newRestaurant)
-        {
-            restaurants.Add(newRestaurant);
-            newRestaurant.Id = restaurants.Max(r => r.Id) + 1;
-            return newRestaurant;
-        }
     }
 }
